Remove the libsimpleservo C API
This is unused and unmaintained. It also added a bit of complication
to the build.
