--- ports/libsimpleservo/capi/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "simpleservo_capi"
-version = "0.0.1"
-authors = ["The Servo Project Developers"]
-license = "MPL-2.0"
-edition = "2018"
-publish = false
-
-[lib]
-name = "simpleservo"
-crate-type = ["cdylib"]
-test = false
-bench = false
-
-[dependencies]
-backtrace = { workspace = true }
-env_logger = { workspace = true }
-lazy_static = { workspace = true }
-log = { workspace = true }
-simpleservo = { path = "../api" }
-surfman = { workspace = true }
-keyboard-types = { workspace = true }
-
-[target.'cfg(target_os = "windows")'.dependencies]
-libc = { workspace = true }
-winapi = { workspace = true, features = ["wingdi", "winuser", "winnt", "winbase", "processenv", "namedpipeapi", "ntdef", "minwindef", "handleapi", "debugapi"] }
-
-[build-dependencies]
-cbindgen = "0.26"
-
-[features]
-debugmozjs = ["simpleservo/debugmozjs"]
-default = ["webdriver", "max_log_level"]
-googlevr = ["simpleservo/googlevr"]
-jitspew = ["simpleservo/jitspew"]
-js_backtrace = ["simpleservo/js_backtrace"]
-max_log_level = ["simpleservo/max_log_level"]
-media-gstreamer = ["simpleservo/media-gstreamer"]
-native-bluetooth = ["simpleservo/native-bluetooth"]
-no-wgl = ["simpleservo/no-wgl"]
-profilemozjs = ["simpleservo/profilemozjs"]
-refcell_backtrace = ["simpleservo/refcell_backtrace"]
-webdriver = ["simpleservo/webdriver"]
-webgl_backtrace = ["simpleservo/webgl_backtrace"]
-xr-profile = ["simpleservo/xr-profile"]
