--- ports/libsimpleservo/capi/cbindgen.toml
@@ -1,8 +1,0 @@
-language = "C"
-include_guard = "simpleservo_h"
-cpp_compat = true
-tab_width = 4
-documentation_style = "c99"
-
-[export]
-exclude = ["OutputDebugStringA"]
