--- ports/libsimpleservo/jniapi/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "simpleservo_jniapi"
-version = "0.0.1"
-authors = ["The Servo Project Developers"]
-license = "MPL-2.0"
-edition = "2018"
-publish = false
-
-[lib]
-name = "simpleservo"
-crate-type = ["cdylib"]
-test = false
-bench = false
-
-[dependencies]
-android_logger = "0.13"
-jni = "0.18.0"
-libc = { workspace = true }
-log = { workspace = true }
-serde_json = { workspace = true }
-# TODO: Once the native-bluetooth feature works for
-# Android, remove the explicit feature list here.
-simpleservo = { path = "../api", default-features = false, features = ["max_log_level", "webdriver"] }
-
-[build-dependencies]
-cc = "1.0"
-
-[features]
-debugmozjs = ["simpleservo/debugmozjs"]
-default = ["max_log_level", "webdriver", "no_static_freetype"]
-googlevr = ["simpleservo/googlevr"]
-js_backtrace = ["simpleservo/js_backtrace"]
-max_log_level = ["simpleservo/max_log_level"]
-media-gstreamer = ["simpleservo/media-gstreamer"]
-native-bluetooth = ["simpleservo/native-bluetooth"]
-webdriver = ["simpleservo/webdriver"]
-webgl_backtrace = ["simpleservo/webgl_backtrace"]
-no_static_freetype = ["simpleservo/no_static_freetype"]
